Move classes to individual files
