+++ Assets/Packages/ServiceLocator/ServiceManagerLogType.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fb4953cf0aee48f5aac83ed5d5e7d45b
+timeCreated: 1594899275
